# perf(toml): Allow opting in to faster hasher