--- a/crates/benchmarks/Cargo.toml
+++ b/crates/benchmarks/Cargo.toml
@@ -13,6 +13,7 @@ default = []
 simd = ["toml_parse/simd"]
 unsafe = ["toml_parse/unsafe", "toml_edit/unsafe"]
 preserve_order = ["toml_old/preserve_order", "toml/preserve_order"]
+fast_hash = ["toml/fast_hash"]
 
 [dependencies]
 serde = { version = "1.0.197", features = ["derive"] }