--- a/crates/toml/Cargo.toml
+++ b/crates/toml/Cargo.toml
@@ -33,6 +33,7 @@ std = ["indexmap?/std"]
 parse = ["dep:toml_parse", "dep:winnow"]
 display = ["std", "dep:toml_edit", "toml_edit?/display"]
 unsafe = ["toml_parse?/unsafe"]
+fast_hash = ["preserve_order", "dep:foldhash"]
 debug = ["std", "toml_parse?/debug", "dep:anstream", "dep:anstyle"]
 
 # Provide a method disable_recursion_limit to parse arbitrarily deep structures
@@ -57,6 +58,7 @@ anstyle = { version = "1.0.8", optional = true }
 toml_edit = { version = "0.22.27", path = "../toml_edit", default-features = false, features = ["serde"], optional = true }
 toml_datetime = { version = "0.6.11", path = "../toml_datetime", default-features = false, features = ["serde", "alloc"] }
 serde_spanned = { version = "0.6.9", path = "../serde_spanned", default-features = false, features = ["serde", "alloc"] }
+foldhash = { version = "0.1.5", default-features = false, optional = true }
 
 [dev-dependencies]
 serde = { version = "1.0.199", features = ["derive"] }