I chose `foldhash` because it is the default of `hashbrown@0.15` (replacing ahash), see rust-lang/hashbrown#563.

Currently, `ahash` is the only hasher I noticed inside of Cargo and that is coming in from `hashbrown@0.14`.

In the future, we could consider allowing the use of a `FixedState` so we could offer `IndexMap` on `alloc`.

```
$ cargo bench --bench 0-cargo -Fpreserve_order -- 0_cargo::toml::document::2-features
   Compiling toml v0.5.11
   Compiling toml v0.8.23 (/home/epage/src/personal/toml/crates/toml)
   Compiling toml_benchmarks v0.0.0 (/home/epage/src/personal/toml/crates/benchmarks)
    Finished `bench` profile [optimized] target(s) in 19.13s
     Running benches/0-cargo.rs (/home/epage/src/personal/toml/target/release/deps/0_cargo-09d7e00849b1b878)
Timer precision: 11 ns
0_cargo              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ toml                            │               │               │               │         │
   ╰─ document                     │               │               │               │         │
      ╰─ 2-features  468.8 µs      │ 1.016 ms      │ 513.5 µs      │ 543.5 µs      │ 100     │ 100


$ cargo bench --bench 0-cargo -Ffast_hash -- 0_cargo::toml::document::2-features
    Finished `bench` profile [optimized] target(s) in 0.02s
     Running benches/0-cargo.rs (/home/epage/src/personal/toml/target/release/deps/0_cargo-ea11bd9a52599c35)
Timer precision: 18 ns
0_cargo              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ toml                            │               │               │               │         │
   ╰─ document                     │               │               │               │         │
      ╰─ 2-features  431.4 µs      │ 870.3 µs      │ 446.2 µs      │ 459.9 µs      │ 100     │ 100
```